requirements.txt
aiohttp==3.8.5
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
playwright==1.39.0
Pillow==10.1.0 